+++ torch/_inductor/bounds.py
@@ -9,7 +9,7 @@ from sympy import Expr
 import torch
 from torch.utils._sympy.value_ranges import bound_sympy, ValueRangeAnalysis, ValueRanges
 
-from .ir import InterpreterShim, LoopBody, LoopBodyBlock
+from .loop_body import InterpreterShim, LoopBody, LoopBodyBlock
 from .utils import cache_on_self, dominated_nodes
 from .virtualized import V
