+++ torch/_inductor/optimize_indexing.py
@@ -6,7 +6,7 @@ import sympy
 import torch
 from torch.utils._sympy.value_ranges import ValueRanges
 
-from .ir import LoopBody
+from .loop_body import LoopBody
 from .utils import dominated_nodes
